Player Movement Script Started
I have added some minor stuf to the scene: a plane, a player object(placeholder) and a main camera. I have created a folder for holding placeholder materials and a folder for holding scripts.

Inside the Player Scripts folder I have started a player movement script. This script is basic, just using the Input Axis to move on a horizonatal plane. There is a Acceleration variable, a max speed variable and a jump variable.

I have it set up so that an accelerating force will NOT be applied to the player if they go over the max speed. I also set up a fallback so that if they do manage to go over the max speed, their speed is set to the max speed..