Primitive Slope Detection
So I managed to get the slope detection to work. I followed this tutorial on the subject (https://www.youtube.com/watch?v=98MBwtZU2kk). A raycast is emitted towards the ground and is used to alter a custom vector for the forward direction. The forward and backward movement now uses this custom forward vector to move around. I have not done horizontal movement on the slope yet, I still need to work that out.